LayoutTests/fast/canvas/webgl/tex-image-and-sub-image-2d-with-image-data-expected.txt

 
Verify texImage2D and texSubImage2D code paths taking ImageData

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing texImage2D with flipY=true and premultiplyAlpha=false
Checking bottom pixel
PASS pixel is correctColor
Checking top pixel
PASS pixel is correctColor
Testing texImage2D with flipY=false and premultiplyAlpha=false
Checking bottom pixel
PASS pixel is correctColor
Checking top pixel
PASS pixel is correctColor
Testing texImage2D with flipY=true and premultiplyAlpha=true
Checking bottom pixel
PASS pixel is correctColor
Checking top pixel
PASS pixel is correctColor
Testing texImage2D with flipY=false and premultiplyAlpha=true
Checking bottom pixel
PASS pixel is correctColor
Checking top pixel
PASS pixel is correctColor
Testing texSubImage2D with flipY=true and premultiplyAlpha=false
Checking bottom pixel
PASS pixel is correctColor
Checking top pixel
PASS pixel is correctColor
Testing texSubImage2D with flipY=false and premultiplyAlpha=false
Checking bottom pixel
PASS pixel is correctColor
Checking top pixel
PASS pixel is correctColor
Testing texSubImage2D with flipY=true and premultiplyAlpha=true
Checking bottom pixel
PASS pixel is correctColor
Checking top pixel
PASS pixel is correctColor
Testing texSubImage2D with flipY=false and premultiplyAlpha=true
Checking bottom pixel
PASS pixel is correctColor
Checking top pixel
PASS pixel is correctColor
PASS successfullyParsed is true

TEST COMPLETE
